The re-write is going awesomely.
There are obviously bugs with it though.

+++ greenies.pl
@@ -1,6 +1,6 @@
 #!/usr/bin/perl
 # vi:tw=0:
-# $Id: greenies.pl 45.1009.XDA0VyAx9t5MalutcbQ2KNABSL0 2007-05-07 06:38:15 -0400 $
+# $Id: greenies.pl 66.1445.lhgFHt8XveWYOoIfdRw3V9/Ij+c 2007-05-07 09:17:05 -0400 $
 
 use strict;
 use GD;
@@ -8,6 +8,7 @@ use GD;
 my $in_set = 0;
 
 while(<STDIN>) {
+    my $drew = 0;
 
     if( m/SET/ ) {
         $in_set ++;
@@ -16,6 +17,7 @@ while(<STDIN>) {
 
     my $image = GD::Image->new("marked.png");
     my $green = $image->colorAllocate(0, 200, 0);
+    my $red   = $image->colorAllocate(190, 0, 0);
     my $black = $image->colorAllocate(0, 0, 0);
 
     REDO: {
@@ -25,21 +27,40 @@ while(<STDIN>) {
 
             $image->line( @lhs=>@rhs, $green );
             $image->filledEllipse(@lhs, 5,5, $green);
-            $image->filledEllipse(@rhs, 6,5, $green);
+            $image->filledEllipse(@rhs, 5,5, $green);
             $image->ellipse(@lhs, 5,5, $black);
             $image->ellipse(@rhs, 5,5, $black);
+
+            $drew ++;
+        }
+
+        while( m/<(\d+)[,\s]+(\d+)>/g ) {
+            my @p = map {$_*23+12} ($1, $2);
+
+            $image->filledEllipse(@p, 9,9, $red);
+            $image->ellipse(@p, 9,9, $black);
+
+            $drew ++;
         }
 
         if( $in_set ) {
             $_ = <STDIN>;
-            redo unless m/DONE/;
+
+            if( m/DONE/ ) {
+                $in_set = 0;
+
+            } else {
+                redo;
+            }
         }
     }
 
-    open my $out, ">", "marked_greenies.png" or die $!;
-    print $out $image->png;
-    close $out;
+    if( $drew ) {
+        open my $out, ">", "marked_greenies.png" or die $!;
+        print $out $image->png;
+        close $out;
 
-    system qw(xv marked_greenies.png);
-    unlink "marked_greenies.png" or die $!;
+        system qw(xv marked_greenies.png);
+        unlink "marked_greenies.png" or die $!;
+    }
 }
